Remove mocha and promises from the public API
Add `configure` and `run` functions to execute the validation suite.

Add `defineSuite` function that can be used to include the validation
suite in an existing Mocha test suite.
