+++ package.json
@@ -2,7 +2,7 @@
   "name": "registry-validator",
   "version": "0.1.0",
   "description": "Validate your npm registry implementation, check if it correctly implements the npm protocol.",
-  "main": "lib/validate.js",
+  "main": "index.js",
   "bin": {
     "registry-validator": "./bin/registry-validator.js"
   },
@@ -34,19 +34,15 @@
     "chai": "^1.9.0",
     "debug": "^0.7.4",
     "fs-extra": "^0.8.1",
+    "mocha": "^1.17.0",
+    "mocha-as-promised": "^2.0.0",
     "npm": "^1.4.4",
     "promise": "^4.0.0",
     "superagent": "^0.17.0",
     "superagent-promise": "^0.1.0"
   },
-  "peerDependencies": {
-    "mocha": "^1.17.0",
-    "mocha-as-promised": "^2.0.0"
-  },
   "devDependencies": {
     "couchapp": "^0.10.0",
-    "mocha": "^1.17.0",
-    "mocha-as-promised": "^2.0.0",
     "npmjs.org": "^1.2.0"
   }
 }
